[SPARK-36352][SQL][3.0] Spark should check result plan's output schema name
